+++ libc/mem.h
@@ -5,7 +5,7 @@
 
 #include "../cpu/type.h"
 
-void memory_copy(uint8_t *source, uint8_t *dest, int nbytes);
+void memory_copy(uint8_t *source, uint8_t *dest, size_t nbytes);
 void memory_set(uint8_t *dest, uint8_t val, uint32_t len);
 
 uint32_t kmalloc(size_t size, int align, uint32_t *phys_addr);
